backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==4.2.17
django-bootstrap5==23.4
django-browser-reload==1.17.0
django-cors-headers==4.6.0
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
regex==2024.11.6
sqlparse==0.5.3
typing_extensions==4.12.2
